First cut- use a delegate strategy to move the launch config select dialog into LutrisWindow.
